finished the assignment
Initialized two variables that are given values by two functions.

The first one takes input from the user and converts it to numbers and removes the $ symbol. Added some "if" statements incase the user decides to type something other than the intended input. The function ends by returning price.

The other function takes the price variable and reduce it by 10% and returns that value to the reducedPrice variable.

Added some functionality by adding these functions to buttons for the user to interact with using getElementById and addEventListener.

Added the ability to see the current value of current price and current reduced price in the html just for fun.