Incorporated the wordGame assignent into my main JavaScript assignment Repo instead
A browser typing game where a random word is shown and the player types it into an input. Correct word gives +1 score and +5 seconds, then given a new word to type.

Functions:
-addWordToDOM
-updateScore
-updateTime
-gameOver

All DOM selections and listeners are attached inside DOMContentLoaded so the scripts runs  when the elements exist.

Given more time I would add functionality to the difficulty settings to affect time given to type words; easy 15s, medium 10s, hard 5s.

and a button to initialize the game instead of directly on page load.